refactor: camel case route functions
