+++ src/http/routes/index.ts
@@ -1,15 +1,15 @@
 import { FastifyInstance } from 'fastify';
 
-import { matchesRoutes } from './matches.routes';
-import { PoolRoutes } from './pools.routes';
-import { PredictionsRoutes } from './predictions.routes';
-import { tournamentsRoutes } from './tournaments.routes';
-import { UserRoutes } from './user.routes';
+import { matchesRoutes } from '@/http/routes/matches.routes';
+import { poolRoutes } from '@/http/routes/pools.routes';
+import { predictionsRoutes } from '@/http/routes/predictions.routes';
+import { tournamentsRoutes } from '@/http/routes/tournaments.routes';
+import { userRoutes } from '@/http/routes/user.routes';
 
 export function routes(app: FastifyInstance): void {
-  app.register(UserRoutes);
+  app.register(userRoutes);
   app.register(matchesRoutes);
-  app.register(PoolRoutes);
-  app.register(PredictionsRoutes);
+  app.register(poolRoutes);
+  app.register(predictionsRoutes);
   app.register(tournamentsRoutes);
 }
